Needed to integrate revised copy with damaged
I created a description of the problem, fix and procedures on what I
did to fix. Also, instructions on how the file must be run from the
WHITE .XCWORKSPACE FILE in order to run properly. This commit is
replacing the revised (fixed version) with my older one with the error.